requirements/base.txt
cython
cpython
PyQt5
PyQt5_tools
fbs
nemo_toolkit
nemo_asr